Implement global CactusBot packet specification
 - Message.Text can not be filled, as described in the FIXME comment
(schemas/schemas.go). This is critical to fix!
 - Packets other than Message are not currently supported, as I was not
able to figure out how to make that work with Go's type system. This
would be a nice feature to have, though, so custom commands could
potentially return other packets (for example, triggering a ban after a
custom warning or something.)
 - util/jsonapi.go was changed to allow for the standard placement of
JSON-API tags (e.g. "attr,foo" instead of "foo,attr"). Is this an issue,
@RPiAwesomeness?
 - createdAt and id are not being populated, but I suspect this has
simply not been completed yet.